lower level reorganization
