+++ requirements.txt
@@ -1,4 +1,3 @@
-numpy==1.21.4
 pandas==1.3.5
 PyMySQL==1.0.2
 python-dateutil==2.8.2
